Desktop/exotel-resume-screener-upload/requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
pdfplumber>=0.11.0
python-docx>=1.1.0
openpyxl>=3.1.0
pdf2image>=1.17.0
pytesseract>=0.3.10
